+++ CHANGELOG.md
@@ -6,11 +6,12 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Finnish translation (thanks Åzze, Olli, Ricky Tigg, Otso Karttunen)
+- Collection: long press on copy/move/rating/tag quick action for quicker action
 - Widget: more shape options
 - Viewer: drag and drop in multi-window mode
 - Vaults: allow hiding like other albums
 - 16 KB memory page size support
+- Finnish translation (thanks Åzze, Olli, Ricky Tigg, Otso Karttunen)
 
 ### Changed